キーコード、配列を少し修正

+++ mysplit/keymap.h
@@ -7,8 +7,8 @@
 #define COL_NUM 6
 #define COL_NUM_2 COL_NUM*2
 
-extern const byte keyMap[ROW_NUM][COL_NUM_2];
-extern const byte keyMapUpper[ROW_NUM][COL_NUM_2];
-extern const byte keyMapLower[ROW_NUM][COL_NUM_2];
+extern const int keyMap[ROW_NUM][COL_NUM_2];
+extern const int keyMapUpper[ROW_NUM][COL_NUM_2];
+extern const int keyMapLower[ROW_NUM][COL_NUM_2];
 
 #endif
